# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Emergent API
  version: 0.1.0
  description: Next.js routes backed by OpenAlex, cached in Postgres/Supabase. See PRD section 16.
servers:
  - url: http://localhost:3000
paths:
...
  /api/ai/reading-path:
    post:
      summary: AI reading path for a topic (auth required, cached 7 days)
      responses:
        "200": { description: Ordered path with why-read notes }
        "401": { description: Sign in required }
        "501": { description: AI not configured }